third_party/blink/web_tests/http/tests/devtools/components/progress-bar-expected.txt

Tests inspector's composite progress bar.


Running: testOneSubProgress
Testing CompositeProgress with a single subprogress:
progress: `undefined' 0 out of 1 done.
progress: `cuckooing' 0 out of 1 done.
progress: `cuckooing' 0.1 out of 1 done.
progress: `meowing' 0.2 out of 1 done.
progress indicator: done
progress: `meowing' 1 out of 1 done.

Running: testMultipleSubProgresses
Testing CompositeProgress with multiple subprogresses:
progress: `undefined' 0 out of 1 done.
progress: `cuckooing' 0.05 out of 1 done.
progress: `cuckooing' 0.05 out of 1 done.
progress: `barking' 0.275 out of 1 done.
progress: `meowing' 0.5 out of 1 done.
progress: `meowing' 0.875 out of 1 done.
progress indicator: done
progress: `meowing' 1 out of 1 done.

Running: testCancel
Testing isCanceled:

Running: testNested
Testing nested subprogresses:
progress: `undefined' 0 out of 1 done.
progress: `undefined' 0 out of 1 done.
progress: `undefined' 0.1 out of 1 done.
progress: `meowing' 0.5 out of 1 done.
progress indicator: done
progress: `meowing' 1 out of 1 done.
